Added pipes and filters example from IDDD
A file with 20k phone numbers is reduced to the number of phone numbers which contain the digit 303. Each processing step is implemented by a filter and piped to another processing component through an event bus. An event store listens to all events and stores them. A projection is used to get the result.

This example was introduced in `Implementing Domain Driven Desig`