Added without directive
By default all parameters of an event are assumed to be payload. This
changes adds the directive 'without' which can be used to exclude
parameters from the payload.

The code is refactored and classes Meta and Payload are removed.
Instead an instance of ParameterContainer is used for meta and payload.
It captures the explicit mappings of parameter names to values and
keeps track of the parameters which are excluded.